Move the power filter to custom filter file and change the imports in files
